requirements.txt
flask
flask-cors
torch
torchvision
pillow
streamlit
